--- ol-cli/src/daemon.rs
@@ -1,6 +1,0 @@
-
-// spawn child web server, which in turn spawns check_runner
-
-// awaits val-wizard to exit, before continuing on loop.
-
-//
