+++ ol-cli/src/server.bak
@@ -1,6 +1,6 @@
 //! web-monitor
 
-use crate::{account, chain_info, check_runner, node_health::{self, NodeHealth}};
+use crate::{node::chain_info, check::{items::Items, runner::run_checks}, node::account, node::{account::AccountInfo, node_health::{self, NodeHealth}}};
 use cli::libra_client::LibraClient;
 use futures::StreamExt;
 use libra_json_rpc_client::AccountAddress;
@@ -14,7 +14,7 @@ use serde_json::json;
 
 // TODO: does this need to be a separate function?
 // create server-sent event
-fn sse_check(info: node_health::Items) -> Result<impl ServerSentEvent, Infallible> {
+fn sse_check(info: Items) -> Result<impl ServerSentEvent, Infallible> {
     Ok(warp::sse::json(info))
 }
 
@@ -36,7 +36,7 @@ pub async fn start_server(node: NodeHealth) {
     // TODO: Perhaps a better way to keep the check cache fresh?
     
     thread::spawn(move || {
-        check_runner::mon(node, true, false);
+        run_checks(node, true, false);
     });
 
     //GET check/ (json api for check data)
@@ -44,7 +44,7 @@ pub async fn start_server(node: NodeHealth) {
         // let mut health = node_health::NodeHealth::new();
         // create server event source from Check object
         let event_stream = interval(Duration::from_secs(10)).map(move |_| {
-            let items = node_health::Items::read_cache().unwrap();
+            let items = Items::read_cache().unwrap();
             // let items = health.refresh_checks();
             sse_check(items)
         });
@@ -103,7 +103,7 @@ pub async fn start_server(node: NodeHealth) {
     let account = warp::path("account").and(warp::get()).map(|| {
         // create server event source
         let event_stream = interval(Duration::from_secs(60)).map(move |_| {
-            let info = crate::account::AccountInfo::read_account_info_cache();
+            let info = AccountInfo::read_account_info_cache();
             sse_account_info(info)
         });
         // reply using server-sent events
